moved SVGs from Navigation catalog to images catalog
